Move helper functions to create response object to the model_response class
