+++ classes/catcontext.php
@@ -220,94 +220,6 @@ class catcontext {
     }
 
     /**
-     * Create response from DB.
-     *
-     * @param int $contextid
-     * @param array $catscaleids
-     * @param int|null $testitemid
-     * @param int|null $userid
-     *
-     * @return array
-     *
-     */
-    public static function getresponsedatafromdb(
-            int $contextid,
-            array $catscaleids,
-            ?int $testitemid = null,
-            ?int $userid = null): array {
-        global $DB;
-
-        list ($sql, $params) = catquiz::get_sql_for_model_input($contextid, $catscaleids, $testitemid, $userid);
-        $data = $DB->get_records_sql($sql, $params);
-        $inputdata = self::db_to_modelinput($data);
-        return $inputdata;
-    }
-
-    /**
-     * Returns data in the following format
-     *
-     * "1" => Array( //userid
-     *     "comp1" => Array( // component
-     *         "1" => Array( //questionid
-     *             "fraction" => 0,
-     *             "max_fraction" => 1,
-     *             "min_fraction" => 0,
-     *             "qtype" => "truefalse",
-     *             "timestamp" => 1646955326
-     *         ),
-     *         "2" => Array(
-     *             "fraction" => 0,
-     *             "max_fraction" => 1,
-     *             "min_fraction" => 0,
-     *             "qtype" => "truefalse",
-     *             "timestamp" => 1646955332
-     *         ),
-     *         "3" => Array(
-     *             "fraction" => 1,
-     *             "max_fraction" => 1,
-     *             "min_fraction" => 0,
-     *             "qtype" => "truefalse",
-     *             "timestamp" => 1646955338
-     *
-     * @param mixed $data
-     *
-     * @return array
-     */
-    private static function db_to_modelinput($data): array {
-        $modelinput = [];
-        // Check: use only most recent answer for each question.
-
-        foreach ($data as $row) {
-            if ($row->state === 'gaveup') {
-                $row->fraction = 0.0;
-            }
-            $entry = [
-                'fraction' => $row->fraction,
-                'max_fraction' => $row->maxfraction,
-                'min_fraction' => $row->minfraction,
-                'qtype' => $row->qtype,
-                'timestamp' => $row->timecreated,
-                'id' => $row->id,
-            ];
-
-            if (!array_key_exists($row->userid, $modelinput)) {
-                $modelinput[$row->userid] = ["component" => []];
-            }
-
-            if (! array_key_exists($row->questionid, $modelinput[$row->userid]['component'])) {
-                $modelinput[$row->userid]['component'][$row->questionid] = $entry;
-                continue;
-            }
-
-            // If we are here, there is already an entry. Only update it if this answer is newer than the last one.
-            if ($row->id > $modelinput[$row->userid]['component'][$row->questionid]['id']) {
-                $modelinput[$row->userid]['component'][$row->questionid] = $entry;
-            }
-        }
-        return $modelinput;
-    }
-
-    /**
      * Save or update catcontext class.
      *
      * @param ?stdClass $newrecord
@@ -414,12 +326,11 @@ class catcontext {
      *
      */
     public function get_strategy(int $catscaleid): model_strategy {
-        $catscaleids = [$catscaleid, ...catscale::get_subscale_ids($catscaleid)];
-        $responsedata = self::getresponsedatafromdb($this->id, $catscaleids);
-        $responses = model_responses::create_from_array($responsedata);
+        $responses = model_responses::create_for_context($this->id);
         $options = json_decode($this->json, true) ?? [];
         $installedmodels = model_strategy::get_installed_models();
         $olditemparams = [];
+        $catscaleids = [$catscaleid, ...catscale::get_subscale_ids($catscaleid)];
         foreach (array_keys($installedmodels) as $model) {
             $olditemparams[$model] = model_item_param_list::load_from_db($this->id, $model, $catscaleids);
         }
